requirements.txt
Django~=4.1.3
psycopg2
django-environ
PyJWT
djangorestframework
django-cors-headers
